Added "Double Trigger:" to every card with a double trigger.
Double Trigger: As the term implies, it is meant to show players when their card will suddenly split into two for a moment. This hopefully will show new players when their 7-fighter board will swallow the fighter they just added to their deck.
Also, lengthened the descriptions of some cards with "When this card is cast or when it dies" so they'd make sense with the new syntax.
I did not touch cards with charges.